make a common class for converting types called TypeConverter
- TypeConverter should replace all the various type mangling
infrastructure around.

- added some first unit tests for TypeConverter. This means
that the infrastructure for unit tests is in place

+++ src/cd3core/flow.cpp
@@ -188,7 +188,7 @@ void Flow::clear() {
 	}
 }
 
-int Flow::size() {
+size_t Flow::size() {
 	cd3assert(fd.defined, "flow not defined");
 	return fd.size;
 }
